V3-4
Cleaned up and made exo-analysis prominence check loop (hopefully) runable. Will need to check that it is running correctly.
Also added a limit to the prominece loop so that it doesn't try to run to infinity/crash/etc. This is easily switchable to start from a high prominence value and go down, increasing the number of maxima detected.

Added a 'process folder' complete message that should appear at end of code to signal all is done.