V3-8 add prominence + fix Area variable
Added a stand-in value for maximum ratio of maxima/cell area for prominence in line 166

Area variable was undefined in the latest versions in the exo_analysis_general function.
Originially, this was defined, as visible in V1. But 'Area' was removed and the code was split into two separate functions.
This commit tries to fix this 'error' and returns Area to the output of exo_analysis_general.